feat: add useTrackPreview3DInteractions hook for 3D interaction handling
- Implemented elevation, rotation, and tilt drag functionalities for 3D shapes.
- Added state management for elevation and rotation dragging.
- Integrated event listeners for mouse and touch interactions.
- Created utility functions for snapping rotation and calculating ground angles.
- Enhanced user experience with interactive previews for polyline shapes.
